move Tile into framework namespace
