+++ GetDressed/GetDressed.csproj
@@ -44,7 +44,7 @@
     <Compile Include="Framework\LocalConfig.cs" />
     <Compile Include="GetDressed.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Tile.cs" />
+    <Compile Include="Framework\Tile.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="manifest.json">
